Format code with Clion, make random device static, print container as a 'generic' function and improve random element selection

+++ 01-exhaustive-search/01-exhaustive-search.cpp
@@ -1,7 +1,7 @@
 #include "01-exhaustive-search.h"
 
 void selection_sort(std::vector<int> &vector) {
-    const int vector_size = (int)vector.size();
+    const int vector_size = (int) vector.size();
     int min_index, tmp;
     for (int i = 0; i < vector_size - 1; ++i) {
         min_index = i;
@@ -19,14 +19,14 @@ void selection_sort(std::vector<int> &vector) {
 
 // Based on SequentialSearch2 algorithm presented in section 3.2 of
 // Introduction to the Design and Analysis of Algorithms (3rd Edition) - Anany Levitin
-int sequential_search(std::vector<int> vector, int key){
-    int n=vector.size();
+int sequential_search(std::vector<int> vector, int key) {
+    int n = (int) vector.size();
     vector.push_back(key);
-    int i=0;
-    while(vector[i] != key){
+    int i = 0;
+    while (vector[i] != key) {
         ++i;
     }
-    if(i<n)
+    if (i < n)
         return i;
     else
         return -1;
